+++ packages/shared/src/utilities/index.ts
@@ -1,8 +1,5 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './color';
-export * from './getBackgroundColorStyles';
-export * from './getBorderStyles';
-export * from './getRadiusStyles';
 export * from './moveItemInArray';
 export * from './react';
