+++ packages/shared/src/utilities/react/getBackgroundColorStyles.spec.ts
