+++ packages/shared/src/utilities/react/getBackgroundColorStyles.ts
@@ -4,7 +4,7 @@ import { CSSProperties } from 'react';
 
 import { Color } from '@frontify/fondue';
 
-import { toRgbaString } from './color';
+import { toRgbaString } from '../color';
 
 export const getBackgroundColorStyles = (backgroundColor: Color): CSSProperties => ({
     backgroundColor: toRgbaString(backgroundColor),
