+++ packages/shared/src/utilities/react/getRadiusStyles.ts
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { CSSProperties } from 'react';
-import { Radius, radiusStyleMap } from '../settings/types';
+import { Radius, radiusStyleMap } from '../../settings/types';
 
 export const getRadiusStyles = (radiusChoice: Radius, hasRadius = false, radiusValue?: number): CSSProperties => ({
     borderRadius: hasRadius ? radiusValue : radiusStyleMap[radiusChoice],
