+++ packages/shared/src/utilities/react/index.ts
@@ -1,3 +1,6 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
+export * from './getBackgroundColorStyles';
+export * from './getBorderStyles';
+export * from './getRadiusStyles';
 export * from './joinClassNames';
